+++ Project/Loothub/src/main/java/com/company/springboot/repository/IItemStatusRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface IItemStatusRepository extends JpaRepository<ItemStatus,Long> {
+public interface IItemStatusRepository extends JpaRepository<ItemStatus,Integer> {
     
 }
