+++ Project/Loothub/src/main/java/com/company/springboot/services/IItemStatusServise.java
@@ -7,8 +7,8 @@ import java.util.List;
 public interface IItemStatusServise {
     public List<ItemStatus> listAll();
     public ItemStatus save(ItemStatus itemStatus);
-    public ItemStatus get(Long id);
-    public void delete(Long id);
+    public ItemStatus get(int id);
+    public void delete(int id);
     public void update(ItemStatus itemStatus);
